Rename `version_info` -> `version-info`
