+++ packages/database/src/cs/collections.ts
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 import { type CrossSectionInfo } from "@lxcat/schema/process";
-import { VersionInfo } from "../shared/types/version_info.js";
+import { VersionInfo } from "../shared/types/version-info.js";
 
 export type CrossSection = {
   versionInfo: VersionInfo;
